:truck: Separate use cases by type
Category use cases package created
